backend/app/services/parser.py
import os
from google.cloud import documentai_v1 as documentai
from google.api_core.client_options import ClientOptions

class DocumentParser:
    """
    Uses Google Document AI to extract structural metadata (Table of Contents, Headers)
    from textbooks to pre-populate the 'Topics' cache.
    """
    def __init__(self, project_id: str, location: str = "us", processor_id: str = None):
        self.project_id = project_id
        self.location = location
        self.processor_id = processor_id or os.getenv("DOCAI_PROCESSOR_ID") # Configurable
        
        self.client = None
        if os.getenv("GOOGLE_APPLICATION_CREDENTIALS") or os.getenv("GOOGLE_CLOUD_PROJECT"):
            try:
                opts = ClientOptions(api_endpoint=f"{location}-documentai.googleapis.com")
                self.client = documentai.DocumentProcessorServiceClient(client_options=opts)
            except Exception as e:
                print(f"⚠️ DocAI Init Warning: {e}")

    def extract_hierarchy(self, gcs_uri: str) -> dict:
        """
        Simulates Document AI extracting a structured TOC.
        Returns: {
            "book_id": "TN_SCERT_PHY_12",
            "chapters": [...]
        }
        """
        print(f"📄 Parsing TOC Hierarchy from {gcs_uri}...")
        
        # Mocking the output of a smart Document AI processor
        return {
            "book_id": "TN_SCERT_PHY_12",
            "title": "Physics - Grade 12 (Volume 1)",
            "chapters": [
                {
                    "chapter_id": "PHY12_01",
                    "title": "Electrostatics",
                    "topics": [
                         {"topic_id": "PHY12_01_01", "title": "Introduction to Electrostatics"},
                         {"topic_id": "PHY12_01_02", "title": "Coulomb's Law"},
                         {"topic_id": "PHY12_01_03", "title": "Electric Field Lines"},
                         {"topic_id": "PHY12_01_04", "title": "Electric Dipole"}
                    ]
                },
                {
                    "chapter_id": "PHY12_02",
                    "title": "Current Electricity",
                    "topics": [
                         {"topic_id": "PHY12_02_01", "title": "Electric Current"},
                         {"topic_id": "PHY12_02_02", "title": "Ohm's Law"},
                         {"topic_id": "PHY12_02_03", "title": "Kirchhoff's Rules"}
                    ]
                }
            ]
        }
